feat: erfpacht overview page
