+++ packages/app/package.json
@@ -24,6 +24,7 @@
     "@nl-portal/nl-portal-user-interface": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-intl": "^6.6.2",
     "react-router-dom": "^6.22.1"
   },
   "devDependencies": {
