+++ packages/localization/src/i18n/messages/en-gb.ts
@@ -89,5 +89,8 @@ export const EN_GB_MESSAGES: Messages = {
     "documentsList.title": "Documents",
     "documentsList.fetchError": "There was an error, try again later.",
     "documentsList.empty": "There are no documents.",
+    "theme.erfpacht.listTitle": "Leasehold contracts",
+    "theme.erfpacht.listAdres": "Address",
+    "theme.erfpacht.listContractnummer": "Contract number",
   },
 };
