+++ packages/localization/src/i18n/messages/nl-nl.ts
@@ -105,6 +105,8 @@ export const NL_NL_MESSAGES: Messages = {
     "actionLabels.approachingDeadlineDay": "nog {days} dag",
     "actionLabels.approachingDeadlineDays": "nog {days} dagen",
     "theme.erfpacht.listTitle": "Erfpachtcontracten",
+    "theme.erfpacht.listAdres": "Adres",
+    "theme.erfpacht.listContractnummer": "Contractnummer",
     "tableList.viewAll": "Bekijk alle items ({total})",
     "tableList.fetchError":
       "Er is een fout opgetreden, probeer het later nogmaals.",
