+++ packages/user-interface/src/pages/ThemeOverviewPage.tsx
@@ -6,14 +6,9 @@ import TasksList from "../components/TasksList";
 import {
   Taak,
   Zaak,
-  ContractBeperkt,
   useGetTakenQuery,
   useGetZakenQuery,
-  useGetErfpachtContractenQuery,
 } from "@nl-portal/nl-portal-api";
-import TableList from "../components/TableList";
-import { useOutletContext } from "react-router-dom";
-import { RouterOutletContext } from "../contexts/RouterOutletContext";
 
 interface Props {
   type: string;
@@ -29,7 +24,7 @@ const ThemeOverviewPage = ({
   children,
 }: Props) => {
   const intl = useIntl();
-  const { paths } = useOutletContext<RouterOutletContext>();
+
   const {
     data: tasksData,
     loading: taskLoading,
@@ -40,19 +35,9 @@ const ThemeOverviewPage = ({
     loading: casesLoading,
     error: casesError,
   } = useGetZakenQuery();
-  const {
-    data: contractsData,
-    loading: contractsLoading,
-    error: contractsError,
-  } = useGetErfpachtContractenQuery();
-  const loading = taskLoading || casesLoading || contractsLoading;
+  const loading = taskLoading || casesLoading;
   const tasks = tasksData?.getTaken.content as Taak[] | undefined;
   const cases = casesData?.getZaken.content as Zaak[] | undefined;
-  const contracts = contractsData?.getErfpachtContracten.content as
-    | ContractBeperkt[]
-    | undefined;
-
-  console.log(contracts);
 
   return (
     <PageGrid>
@@ -85,55 +70,6 @@ const ThemeOverviewPage = ({
           }
         />
       )}
-      <TableList
-        loading={loading}
-        error={Boolean(contractsError)}
-        titleTranslationId={`theme.${type}.listTitle`}
-        headers={["Adres", "Contractnummer"]}
-        rows={contracts?.map((contract) => {
-          const href = paths.themeDetails(type, contract.id);
-          const { straatnaam, huisnummer, woonplaats } = contract.adressen[0];
-          return [
-            { children: `${straatnaam} ${huisnummer} ${woonplaats}`, href },
-            { children: contract.id, href },
-          ];
-        })}
-        // rows={[
-        //   [
-        //     {
-        //       children: "Westerstraat 393 Den Haag",
-        //       href: paths.themeDetails(type, "123"),
-        //     },
-        //     {
-        //       children: "‘s-Gravenhage AF 2679",
-        //       href: paths.themeDetails(type, "123"),
-        //     },
-        //     { children: "78435785", href: paths.themeDetails(type, "123") },
-        //   ],
-        //   [
-        //     {
-        //       children: "Westerstraat 393 Den Haag",
-        //       href: paths.themeDetails(type, "123"),
-        //     },
-        //     {
-        //       children: "‘s-Gravenhage AF 2679",
-        //       href: paths.themeDetails(type, "123"),
-        //     },
-        //     { children: "78435785", href: paths.themeDetails(type, "123") },
-        //   ],
-        //   [
-        //     {
-        //       children: "Westerstraat 393 Den Haag",
-        //       href: paths.themeDetails(type, "123"),
-        //     },
-        //     {
-        //       children: "‘s-Gravenhage AF 2679",
-        //       href: paths.themeDetails(type, "123"),
-        //     },
-        //     { children: "78435785", href: paths.themeDetails(type, "123") },
-        //   ],
-        // ]}
-      />
       {children}
     </PageGrid>
   );
